Ex3 done.

+++ Matchismo/Model/Card.h
@@ -16,7 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, getter=isChosen) BOOL chosen;
 @property (nonatomic, getter=isMatched) BOOL matched;
 
--(int)match:(NSArray *)otherCard;
+-(NSInteger)match:(NSArray *)otherCard;
 
 @end
